+++ src/agent/index.ts
@@ -3,8 +3,8 @@
 export { setRuntime, Runtime } from './runtime';
 export { setSyncStateProvider, SyncState } from './commands/connection';
 export { dispatchAgentCommand } from './dispatcher';
-export { AGENT_API_VERSION, getPortFilePath } from './portFile';
-export { startAgentHttpServer, stopAgentHttpServer, HttpServerState } from './transport/http';
+export { AGENT_API_VERSION, getPortFilePath, setGlobalPortFileEnabled } from './portFile';
+export { startAgentHttpServer, stopAgentHttpServer, HttpServerState, AGENT_CONNECT_SNIPPET } from './transport/http';
 export { startAgentFileTransport, logAgentRequestToFile, FileTransportHandle } from './transport/file';
 export * as pendingRegistry from './pendingRegistry';
 export { AgentError, AgentErrorCode, httpStatusForCode, inferCodeFromMessage } from './errors';